MutationAdmission: WIP tried ClusterIP service type
The issue:
The container runtime must pull from the ClusterIP -> it must trust the
ClusterIP
This requires the use of a privileged daemonset that fetches the
ClusterIP and writes it to /etc/containers/certs.d (or
/etc/containerd/...). This also requires the Trow server to fetch its
service IP in the mutating webhook. Not an elegant solution.
