+++ charts/trow/templates/job-create-secret-tls.yaml
@@ -1,4 +1,3 @@
-{{- if .Values.trow.validation.enabled -}}
 apiVersion: batch/v1
 kind: Job
 metadata:
@@ -9,7 +8,7 @@ metadata:
     "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
   labels:
     {{- include "trow.labels" . | nindent 4 }}
-    app.kubernetes.io/component: admission-webhook
+    app.kubernetes.io/component: registry
 spec:
 {{- if .Capabilities.APIVersions.Has "batch/v1alpha1" }}
   ttlSecondsAfterFinished: 0
@@ -19,7 +18,7 @@ spec:
       name: {{ include "trow.fullname" . }}-admission-create
       labels:
         {{- include "trow.labels" . | nindent 8 }}
-        app.kubernetes.io/component: admission-webhook
+        app.kubernetes.io/component: registry
     spec:
       containers:
         - name: create
@@ -29,7 +28,7 @@ spec:
             - create
             - --host={{ include "trow.fullname" . }}-admission,{{ include "trow.fullname" . }}-admission.$(POD_NAMESPACE).svc
             - --namespace=$(POD_NAMESPACE)
-            - --secret-name={{ include "trow.fullname" . }}-admission
+            - --secret-name={{ include "trow.fullname" . }}-tls
           env:
             - name: POD_NAMESPACE
               valueFrom:
@@ -38,9 +37,8 @@ spec:
           securityContext:
             allowPrivilegeEscalation: false
       restartPolicy: OnFailure
-      serviceAccountName: {{ include "trow.fullname" . }}-admission
+      serviceAccountName: {{ include "trow.fullname" . }}-tls-patch
       securityContext:
         runAsNonRoot: true
         runAsUser: 2000
         fsGroup: 2000
-{{- end }}
