+++ charts/trow/templates/tls/clusterRole.yaml
