+++ charts/trow/templates/tls/serviceaccount.yaml
@@ -2,12 +2,11 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: {{ include "trow.fullname" . }}-admission
+  name: {{ include "trow.fullname" . }}-tls-patch
   namespace: {{ .Release.Namespace }}
   annotations:
     "helm.sh/hook": pre-install,pre-upgrade,post-install,post-upgrade
     "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
   labels:
     {{- include "trow.labels" . | nindent 4 }}
-    app.kubernetes.io/component: admission-webhook
 {{- end }}
