--- charts/trow/templates/webhooks/service.yaml
@@ -1,20 +1,0 @@
-{{ if .Values.trow.validation.enabled }}
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "trow.fullname" . }}-admission
-  labels:
-    {{- include "trow.labels" . | nindent 4 }}
-  namespace: {{ .Release.Namespace }}
-spec:
-  type: ClusterIP
-  ports:
-    - port: 443
-      targetPort: webhook
-      name: https-webhook
-    {{- if semverCompare ">=1.20" .Capabilities.KubeVersion.Version }}
-      appProtocol: https
-    {{- end }}
-  selector:
-    {{- include "trow.selectorLabels" . | nindent 4 }}
-{{- end }}
